requirements.txt
asgiref==3.9.1
diskcache==5.6.3
Django==5.0.1
django-cors-headers==4.3.1
django-filter==25.1
django-geojson==4.2.0
django-leaflet==0.31.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
drf-yasg==1.21.7
inflection==0.5.1
Jinja2==3.1.6

MarkupSafe==3.0.2
numpy==2.2.4
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
shapely==2.0.7
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
